perf(engine): probe cache and matching fast paths, plus crash guards
Performance:
- memoize natural_content_width per (node, available, depth): nested
  auto-width boxes and tables drop from O(2^k) layouts to polynomial
- skip provably no-op flex stretch re-layouts; allocation-free grow
  early-out
- share sibling lists per parent (Rc) during selector matching,
  precompute specificity per sheet, skip ::before/::after scans when
  no sheet uses them, short-circuit hover candidate scans
- binary-search ellipsis truncation, cumulative caret measurement,
  lightweight TextRunStyle instead of full ComputedStyle clones
- z-index-free fast paths in paint and hit-testing

Robustness and correctness:
- MAX_DEPTH (512) guards in style/layout/paint/selection recursion
- clamp font-size to 0..=512px, reject oversized WOFF tables and
  images, usize image offsets, clamp negative widths/heights
- resolve abs-positioned top/bottom percentages against the
  containing block height, not width
- skip floats and out-of-flow children in first-child margin
  collapsing (fixes a debug panic and wrong release layout)
- fallback placement for floats that exhaust placement attempts
- line-height percentages resolve against font-size; srcset skips
  empty candidates; text nodes take vertical-align/transition from
  their parent so sup/sub shifts and hover fades keep working

+++ crates/lumen-engine/src/raster.rs
@@ -1488,7 +1488,7 @@ fn blit_image(
         for pixel_x in x0..x1 {
             let u = ((pixel_x as f32 - rect.x) / rect.width).clamp(0.0, 1.0);
             let source_x = ((u * image.width as f32) as u32).min(image.width - 1);
-            let offset = ((source_y * image.width + source_x) * 4) as usize;
+            let offset = (source_y as usize * image.width as usize + source_x as usize) * 4;
             let [r, g, b, a] = image.rgba[offset..offset + 4] else {
                 continue;
             };
